requirements.txt
Appium-Python-Client~=1.2.0
selenium~=3.141.0
requests~=2.25.1
lxml~=4.6.3